+++ xtask/src/installer.rs
@@ -74,11 +74,15 @@ pub(crate) fn build_installer(flags: BuildInstaller) -> Result<()> {
             "cargo build -p chewing_tip {release...} --target {x86_64_target}"
         )
         .run()?;
-        cmd!(
-            sh,
-            "cargo build -p chewing-preferences {release...} --target {x86_64_target}"
-        )
-        .run()?;
+        {
+            let _p = sh.push_dir("preferences");
+            // FIXME https://github.com/matklad/xshell/issues/82
+            if sh.path_exists("/usr/bin/npm") {
+                cmd!(sh, "npm run tauri -- build --target {x86_64_target}").run()?;
+            } else {
+                cmd!(sh, "npm.cmd run tauri -- build --target {x86_64_target}").run()?;
+            }
+        }
         cmd!(
             sh,
             "cargo build -p chewing-editor {release...} --target {x86_64_target}"
@@ -162,23 +166,28 @@ pub(crate) fn build_installer(flags: BuildInstaller) -> Result<()> {
             "build/installer/x64",
         );
     }
-    for file in [
-        "ChewingPreferences.exe",
-        "chewing-editor.exe",
-        "chewing-update-svc.exe",
-        "tsfreg.exe",
-    ] {
+    sh.copy_file(
+        format!(
+            "preferences/src-tauri/{}/ChewingPreferences.exe",
+            x86_64_target_dir.display()
+        ),
+        "build/installer",
+    )?;
+    // May not exist in cross-compile environment.
+    let _ = sh.copy_file(
+        format!(
+            "preferences/src-tauri/{}/ChewingPreferences.pdb",
+            x86_64_target_dir.display()
+        ),
+        "build/installer",
+    );
+    for file in ["chewing-editor.exe", "chewing-update-svc.exe", "tsfreg.exe"] {
         sh.copy_file(
             format!("{}/{file}", x86_64_target_dir.display()),
             "build/installer",
         )?;
     }
-    for file in [
-        "ChewingPreferences.pdb",
-        "chewing-editor.pdb",
-        "chewing-update-svc.pdb",
-        "tsfreg.pdb",
-    ] {
+    for file in ["chewing-editor.pdb", "chewing-update-svc.pdb", "tsfreg.pdb"] {
         let _ = sh.copy_file(
             format!("{}/{file}", x86_64_target_dir.display()),
             "build/installer",